add: new feature to create and delete title
