+++ src/components/UserProfile/QuickSetupModal/QuickSetupModal.css
@@ -9,7 +9,7 @@
   color: white;
   left: 10px;
   right: 10px;
-  top: 30vh;
+  top: 23vh;
   display: none;
 }
