+++ src/components/UserProfile/QuickSetupModal/QuickSetupModal.jsx
@@ -1,7 +1,9 @@
 import React, { useState, useEffect } from 'react';
 import AddNewTitle from './AddNewTitle';
 import AssignPopUp from './AssignPopUp';
+import QuickSetupCodes from './QuickSetUpCodes';
 import { getAllTitle } from '../../../actions/title';
+
 
 import {
   Button,
@@ -13,53 +15,30 @@ import {
 import styles from './QuickSetupModal.css';
 
 
-
 const QuickSetupModal = (props) => {
   const {
 jobTitle  } = props;
-  const [addTitle, setAddTitle] = useState(false);
+  const [onAddTitle, setAddTitle] = useState(false);
   const [assignPopUp, setAssignPopup] = useState(false);
   const [titles, setTitles] = useState([]);
   const [curtitle, setTitleOnClick] = useState('');
+  const [submittoggler, setSubmit] = useState(false);
 
   useEffect(() => {
     getAllTitle()
       .then(res => {console.log(res.data); setTitles(res.data)})
       .catch(err => console.log(err))
-  },[])
-  const qsc = titles.map((title) =>
-  <div role="button" id="wrapper"     className="role-button bg-warning"
-  onClick={()=> {
-    setAssignPopup(true);
-    setTitleOnClick(title);
-    }}
-    value="Software Engineer">{title?.shortName}
-    <div className="title">
-    <div className="summary">{title?.titleName}</div>
-    </div>
-  </div>
-)
+  },[submittoggler])
 
   return (
     <div className="container pt-3">
-      <div className="blueSquares">
-        {qsc}
-      {/* <div role="button" id="wrapper"  className="role-button bg-warning" onClick={()=> setAssignPopup(true)} value="Software Engineer">SE
-      <div className="title">
-        <div className="summary">Software Engineer</div>
-      </div>
-
-      </div>
-      <div role="button" className="role-button bg-warning" onClick={()=> setAssignPopup(true)} value="Software Engineer">ME</div>
-      <div role="button" className="role-button bg-warning" onClick={()=> setAssignPopup(true)} value="Software Engineer">SE</div>
-      <div role="button" className="role-button bg-warning" onClick={()=> setAssignPopup(true)} value="Software Engineer">SE</div> */}
-
-    </div>
+        <QuickSetupCodes titles={titles} setAssignPopup={setAssignPopup}
+        setTitleOnClick={setTitleOnClick}/>
 
       <div className="col text-center">
-        <Button onClick={(e) => setAddTitle(true)}>Add A New Title</Button>
+        <Button className="mt-5" onClick={(e) => setAddTitle(true)}>Add A New Title</Button>
       </div>
-      <AddNewTitle isOpen={addTitle} setIsOpen={setAddTitle} toggle={setAddTitle} />
+      <AddNewTitle isOpen={onAddTitle} setIsOpen={setAddTitle} toggle={setAddTitle} setSubmit={setSubmit} submittoggler={submittoggler}/>
       <AssignPopUp isOpen={assignPopUp} setIsOpen={setAssignPopup} toggle={setAssignPopup} title={curtitle}/>
     </div>
   )
